+++ includes/entrypoint/init-compose.sh
@@ -8,4 +8,8 @@ FILES=()
 
 FILES+=("-f docker-compose.yml")
 
+if [[ -n "$PROJECT_CONTEXT" ]]; then
+    FILES+=("-f docker-compose.$PROJECT_CONTEXT.yml")
+fi
+
 export DOCKER_COMPOSE="env ${ENV_PREFIX[@]} docker-compose ${FILES[@]}"