+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added dockerfile directives
 - Added proxy for common docker-compose commands such as `up`, `down`, `start`, etc
 - Added nested dependency resolvement during both initialization & build stages
+- Added project configuration (name and context)
 
 ### Updated
 - Upgraded yq to version 4 (removed undocumented dependency on jq)